Update Back Projection (working)
This code has the code I have been working on commented out because at first it produced only a purple image, so I've been trying to make it work to make the program faster. It is currently not functional, but if it is left commented out, a better image is produced.